requirements.txt
Django==5.0.1
djangorestframework==3.14.0
psycopg2-binary==2.9.11
django-cors-headers==4.3.1
python-decouple==3.8
drf-yasg==1.21.7
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
setuptools>=65
wheel>=0.41